requirements.txt
selenium>=4.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
numpy>=1.24.0
flask>=2.3.0